move prompt and lua_runner and fix overlap calculation error
