+++ .gitignore
@@ -1,2 +1,4 @@
 __pycache__/
 *.py[cod]
+*.txt
+!examples/*.txt
